refactor(src): API 타입 분리 및 import 시 'type' 키워드 추가
