+++ frontend/src/hooks/template/useTemplateEditQuery.spec.tsx
@@ -2,7 +2,7 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { renderHook, waitFor } from '@testing-library/react';
 import { PropsWithChildren } from 'react';
 
-import { TemplateEditRequest } from '@/types/template';
+import type { TemplateEditRequest } from '@/types';
 import { useTemplateEditQuery } from './useTemplateEditQuery';
 
 const queryClient = new QueryClient();
