+++ frontend/src/hooks/template/useTemplateQuery.ts
@@ -1,7 +1,7 @@
 import { UseQueryResult, useQuery } from '@tanstack/react-query';
 
 import { QUERY_KEY, getTemplate } from '@/api';
-import { Template } from '@/types/template';
+import type { Template } from '@/types';
 
 export const useTemplateQuery = (id: number): UseQueryResult<Template, Error> =>
   useQuery<Template, Error>({
