+++ frontend/src/types/api.ts
@@ -1,0 +1,21 @@
+import { Category, Snippet, Template } from './template';
+
+export interface TemplateEditRequest {
+  title: string;
+  createSnippets: Snippet[];
+  updateSnippets: Snippet[];
+  deleteSnippetIds: number[];
+}
+
+export interface TemplateListResponse {
+  templates: Template[];
+}
+
+export interface TemplateUploadRequest {
+  title: string;
+  snippets: Snippet[];
+}
+
+export interface CategoryListResponse {
+  categories: Category[];
+}
