+++ frontend/src/types/index.ts
@@ -1,0 +1,2 @@
+export type { Snippet, Tag, Category, Template, TemplateListItem } from './template';
+export type { TemplateEditRequest, TemplateListResponse, TemplateUploadRequest } from './api';
